+++ src/store/reducers/executeQuery.ts
@@ -12,7 +12,7 @@ import type {QueryRequestParams, QueryMode, QuerySyntax} from '../../types/store
 import {QUERIES_HISTORY_KEY} from '../../utils/constants';
 import {QUERY_MODES, QUERY_SYNTAX, parseQueryAPIExecuteResponse} from '../../utils/query';
 import {parseQueryError} from '../../utils/error';
-import {settingsManager} from '../../utils/settings';
+import {settingsManager} from '../../services/settings';
 import '../../services/api';
 
 import {createRequestActionTypes, createApiRequest} from '../utils';
